+++ packages/web3-types/CHANGELOG.md
@@ -42,3 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Web3EthExecutionAPI` export (#5441)
 -   `Web3NetAPI` export (#5441)
 -   `EthPersonalAPI` export (#5441)
+
+### Changed
+
+-   `Web3APISpec`, `Web3APIMethod`, and `Web3APIParams` now supports `unknown` APIs (#5393)